requirements.txt
anyio==4.3.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-utils==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.15
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
python-dotenv~=1.0.1
passlib==1.7.4